Isolate characterization execution from verifier
