+++ .supportability-characterization.json
@@ -6,9 +6,7 @@
         "src/supportability_gate/cli.py",
         "src/supportability_gate/complexity_metrics.py",
         "src/supportability_gate/contract.py",
-        "src/supportability_gate/git_changes.py",
-        "src/supportability_gate/responses_transport.py",
-        "src/supportability_gate/semantic_contract.py"
+        "src/supportability_gate/git_changes.py"
       ],
       "id": "gate-cli",
       "kind": "cli"
