+++ src/supportability_gate/semantic_contract.py
@@ -200,11 +200,7 @@ def request_payload(packet: EvidencePacket) -> dict[str, Any]:
         "separation of concerns for every supplied non-removed Python or frontend source path. "
         "Feasibility means the shown code paths can perform their stated behavior without a "
         "blocking internal defect. Security means identities, secrets, evidence, and trust "
-        "boundaries fail closed and target code cannot execute in the local evaluator or semantic "
-        "verifier. Owner-authorized characterization drivers may execute only through fixed Python "
-        "or Node vectors in separate GitHub-hosted base/head jobs; that isolated execution is "
-        "required, while local, self-hosted, configurable, or arbitrary execution blocks. Runtime "
-        "and protected-merge proof "
+        "boundaries fail closed and target code cannot execute. Runtime and protected-merge proof "
         "is gathered separately and is not a prerequisite for this code verdict. A fresh head "
         "without a trusted verdict is the offline fail-closed case; a completed unchanged exact-"
         "evidence verdict may be replayed. Strict up-to-date branch protection handles later base "
